+++ src/Language/Haskell/TH/Instances/Internal.hs
@@ -6,13 +6,19 @@
 -- transformer data types.
 module Language.Haskell.TH.Instances.Internal
   ( deriveQuasiTrans
+  , deriveQuoteTrans
   , Proxy2
   ) where
 
 import qualified Control.Monad.Trans as MTL (lift)
 import Language.Haskell.TH.Lib
 import Language.Haskell.TH.Ppr (pprint)
-import Language.Haskell.TH.Syntax
+import Language.Haskell.TH.Syntax hiding (newName)
+#if MIN_VERSION_template_haskell(2,17,0)
+import Language.Haskell.TH.Syntax (newName)
+#else
+import Language.Haskell.TH.Syntax.Compat (Quote(newName))
+#endif
 
 deriveQuasiTrans ::
      Q Type  -- ^ The instance head. For example, this might be of the form:
@@ -24,22 +30,8 @@ deriveQuasiTrans ::
              --   use 'Proxy2' as an ugly workaround.
   -> Q Exp   -- ^ The implementation of 'qRecover'
   -> Q [Dec] -- ^ The 'Quasi' instance declaration
-deriveQuasiTrans qInstHead qRecoverExpr = do
-  instHead    <- qInstHead
-  let (instCxt, mangledInstTy) = decomposeType instHead
-      qInstCxt = return instCxt
-      qInstTy  = case mangledInstTy of
-                   ConT proxy2 `AppT` instTy
-                     |  proxy2 == ''Proxy2
-                     -> conT ''Quasi `appT` return instTy
-                   _ -> fail $ "Unexpected type " ++ pprint mangledInstTy
-  instDec <- instanceD qInstCxt qInstTy qInstMethDecs
-  return [instDec]
+deriveQuasiTrans qInstHead qRecoverExpr = deriveTrans ''Quasi qInstHead qInstMethDecs
   where
-    decomposeType :: Type -> (Cxt, Type)
-    decomposeType (ForallT _tvbs ctxt ty) = (ctxt, ty)
-    decomposeType ty                      = ([],   ty)
-
     qInstMethDecs :: [Q Dec]
     qInstMethDecs =
       let instMeths :: [(Name, Q Exp)]
@@ -93,6 +85,41 @@ deriveQuasiTrans qInstHead qRecoverExpr = do
           mkDec methName methRhs = valD (varP methName) (normalB methRhs) []
 
       in map (uncurry mkDec) instMeths
+
+deriveQuoteTrans ::
+     Q Type  -- ^ The instance head. For example, this might be of the form:
+             --
+             --   > [t| forall r m. Quote m => Proxy2 (ReaderT r m) |]
+             --
+             --   Why use 'Proxy2' instead of 'Quote'? Sadly, GHC 7.0/7.2 will
+             --   not accept it if you use the latter due to old TH bugs, so we
+             --   use 'Proxy2' as an ugly workaround.
+  -> Q [Dec] -- ^ The 'Quasi' instance declaration
+deriveQuoteTrans qInstHead = deriveTrans ''Quote qInstHead qInstMethDecs
+  where
+    qInstMethDecs :: [Q Dec]
+    qInstMethDecs = [valD (varP 'newName) (normalB [| MTL.lift . newName |]) []]
+
+deriveTrans ::
+     Name    -- ^ The class to derive
+  -> Q Type  -- ^ The instance head
+  -> [Q Dec] -- ^ The implementations of each method in the instance
+  -> Q [Dec] -- ^ The derived instance declaration
+deriveTrans clsName qInstHead qInstMethDecs = do
+  instHead <- qInstHead
+  let (instCxt, mangledInstTy) = decomposeType instHead
+      qInstCxt = return instCxt
+      qInstTy  = case mangledInstTy of
+                   ConT proxy2 `AppT` instTy
+                     |  proxy2 == ''Proxy2
+                     -> conT clsName `appT` return instTy
+                   _ -> fail $ "Unexpected type " ++ pprint mangledInstTy
+  instDec <- instanceD qInstCxt qInstTy qInstMethDecs
+  return [instDec]
+  where
+    decomposeType :: Type -> (Cxt, Type)
+    decomposeType (ForallT _tvbs ctxt ty) = (ctxt, ty)
+    decomposeType ty                      = ([],   ty)
 
 -- | See the Haddocks for 'deriveQuasiTrans' for an explanation of why this
 -- type needs to exist.